Lets reduce image size by removing contents of /tmp
Keep in mind that in some brands/models we copy the kernel file from /tmp so cleaning the /tmp must be done after that task.

Also some brands/models could remove /boot contents as well but in some cases we need that directory contents for booting the image so I'm not gonna touch it now.